requirements.txt
APScheduler==3.11.0
certifi==2025.1.31
cryptography==3.4.8
Flask==3.1.0
line_bot_sdk==3.14.5
lxml==5.3.0
numpy==2.2.5
openai==1.77.0
Pillow==11.2.1
pytz==2022.1
requests==2.32.3
requests
line-bot-sdk
